feat(mail): file inbound mail with a decision model, not a regex
QUOTE_REQUEST_PATTERN was eleven alternations long and still could not tell
"please quote your best rate" sent to us from "the commercial quotation for
Focus Air" sent by us, because the discriminating fact is who is asking whom.
Every false positive bought another negative lookahead.

Filing is now Jev (typesafe/jev-1.13), a TypeSafe System One model: it takes a
state and typed questions and returns a choice with a probability per option.
It runs on OpenRouter's Decisions API — POST /api/alpha/decisions, NOT under
/api/v1 — so it gets its own client rather than a branch in OpenRouterClient.
A chat-completions body there is a 400 and the reverse is a 404 that reads
like the model is missing. Same OPENROUTER_API_KEY, same bill.

Gemma is untouched. Documents — invoices, packing lists, scans — still go to
google/gemma-4-31b-it through the chat endpoint. Jev only picks a folder.

Only classification moved. The weight, pieces, dimension and lane patterns
stayed: they read a number off a page and those numbers end up on a customs
declaration, and Jev is documented as unreliable on numbers and generates no
values at all.

It sits fourth in the chain — tenant rules, then a domain we already invoice,
then the platform airline/domain directory, THEN the model. The first three
are facts and a reading of prose must not overturn a fact. It is also why a
settled mailbox is cheap: most mail never reaches the model.

The rubric is prose in config/mail_intent.php, so changing how mail is filed
is a diff of English rather than of alternations. rubric_version is stamped on
every decision. Two of its sentences were written by a measured run rather
than guessed at: without "the rate is already agreed and they are now asking
us to book", a client writing "Agreed rate 350++, please share the confirmed
booking schedule" came back at 0.59 and fell below the floor; without "a short
reply that mentions a figure but names no cargo, no route and no request", a
two-word reply reading "about 480 kg" was filed as an enquiry at 0.85 and
would have minted a document number. 13/13 on the acceptance set after both,
US$0.0005 and 573ms average for the run.

Credits learned decimals. The ledger was INTEGER, so every priced thing had to
cost a whole document. A mail is about ₹0.003 against a document's ₹0.025 — a
tenth — which on an integer ledger rounds to 1 (eight times the real cost) or
0 (claiming the AI reading their mail is free). DECIMAL(12,2) throughout, and
every balance is rounded before it is stored so float arithmetic cannot
accumulate in the column.

Classification moved OUT of the per-message transaction. As a regex that was
free; as a hosted model it held a lockForUpdate on the tenant's companies row
— taken to charge the credit — across a 600ms HTTP call, queueing every other
mail for that tenant and every document extraction behind one request. The
trade is the one indexAttachments already makes.

RegexClassificationService is now Mail\MailFilingService. A class called
"Regex…" that calls a hosted model is a sentence everyone has to unlearn.

The per-source accuracy breakdown in Super Admin → Mail filing was built and
then removed on request: if the model files well, the screen is not needed to
find out. The columns behind it are still written, because "which source, at
what confidence, under which rubric" cannot be answered retroactively.

Note: this change was split by accident. JevClient, MailIntentClassifier,
config/mail_intent.php and the migration were swept into cfb4e93e (feat(ports))
by a concurrent session's commit-all.

+++ tests/Feature/CreditGrantTest.php
@@ -17,17 +17,17 @@ class CreditGrantTest extends TestCase
         $company = Company::create(['name' => 'Grant Co', 'code' => 'GRT', 'tier' => 'tactical']);
 
         $this->artisan('credits:grant-monthly')->assertSuccessful();
-        $this->assertSame(500, (int) $company->fresh()->ocr_credits_balance);
+        $this->assertSame(500.0, (float) $company->fresh()->ocr_credits_balance);
 
         // Some used, then a second run the same month changes nothing.
         $company->forceFill(['ocr_credits_balance' => 480])->save();
         $this->artisan('credits:grant-monthly')->assertSuccessful();
-        $this->assertSame(480, (int) $company->fresh()->ocr_credits_balance);
+        $this->assertSame(480.0, (float) $company->fresh()->ocr_credits_balance);
 
         // Moved up to Command mid-month: 2,000 − 500 already granted = 1,500 added.
         $company->forceFill(['tier' => 'command'])->save();
         $this->artisan('credits:grant-monthly')->assertSuccessful();
-        $this->assertSame(1980, (int) $company->fresh()->ocr_credits_balance);
+        $this->assertSame(1980.0, (float) $company->fresh()->ocr_credits_balance);
         $this->assertSame(2000, (int) DB::table('ocr_credit_transactions')->where('company_id', $company->id)->where('transaction_type', 'monthly_grant')->sum('amount'));
     }
 }
